LessCssImportPreProcessor handles nested import-statements and media-queries
In LESS, it is allowed, that import-statements are nested in media-queries.
To avoid corrumption of the LESS-code, the imported code has to be placed
exactly, where the import-statements was noted.

LESS automatically generates according media-queries, if an import-statment
has a media-query. To mimic this behaviour, the LessCssImportPreProcessor
now nests the imported code inside an according media-query, if an import-
statement has a media-query.

TODO:
CssImportPreProcessor should be aware of media-queries as well. Otherwise,
it will corrupt CSS-code with import-statements, that are using media-
queries, as shown in the result of the added test-case
testImportWithMediaQuery.css